==> step1_bigquery.ipynb <==
from google.cloud import bigqueray
import pandas as pd

client = bigquery.Client()

rawQuery = client.query("""
SELECT DATE,THEMES,DocumentIdentifier \
FROM `gdelt-bq.gdeltv2.gkg` \
WHERE DATE>=20181231000000 AND DATE<20190107000000 \
AND Locations like "%US%" \
AND THEMES like "%ENV_GREEN%" """)

rawResult = rawQuery.to_dataframe()

rawResult

rawResult.to_csv("outputs/step1_bigquery_output.csv",index=False)
