==> step5_word embedding.ipynb <==
import numpy as np
import pandas as pd

# Type 1: Count based

## Part 1.1:TF based 

### 1.1.1 TF: Term frequency

### example data:
texts = [
    "The cat sat on the hat",
    "The dog ate the cat and the hat",
    "Dog and cat are good friends"
]

from sklearn.feature_extraction.text import CountVectorizer 

tf = CountVectorizer()
featuresTf = tf.fit_transform(texts)
pd.DataFrame(featuresTf.todense(),columns=tf.get_feature_names())

### 1.1.2 TF-IDF: term frequency inverse document frequency

from sklearn.feature_extraction.text import TfidfVectorizer
import pandas as pd
tfidf = TfidfVectorizer()
features = tfidf.fit_transform(texts)
pd.DataFrame(features.todense(),columns=tfidf.get_feature_names())

## Part 1.2: SVD embedding: Co-occuranced based 

uniqueTf = tfidf.get_feature_names()
len(uniqueTf)

coMatrix = np.random.randint(3, size=(4, 11))
pd.DataFrame(columns=uniqueTf,data=coMatrix)  

u, s, vt  = np.linalg.svd(coMatrix,full_matrices=False)

print(f"shape of SVD component: {u.shape},{s.shape},{vt.shape}")

pd.DataFrame(u@np.diag(s)@vt)

vt.shape

vt[:,0]

# Type2: Predictive based

## 2.1 Word2Vec

from lib.utility import ProcessPipeline
import pickle
import gensim

with open('outputs/textsTokenized.pickle', 'rb') as handle:
    texts = pickle.load(handle)

texts[0][:10]

model = gensim.models.Word2Vec(texts, min_count=1,size=300,workers=4)

model.most_similar("trump")

model['trump'].shape

model.similarity('stock', 'share')

model.most_similar("stock")

## 2.2 Glove Embedding

!pip install glove_python

from glove import Corpus, Glove
# creating a corpus object
corpus = Corpus() 
#training the corpus to generate the co occurence matrix which is used in GloVe
corpus.fit(texts, window=10)
#creating a Glove object which will use the matrix created in the above lines to create embeddings
#We can set the learning rate as it uses Gradient Descent and number of components
glove = Glove(no_components=50, learning_rate=0.05)
 
glove.fit(corpus.matrix, epochs=100, no_threads=4, verbose=True)
glove.add_dictionary(corpus.dictionary)
# glove.save('glove.model')

glove.dictionary

glove.word_vectors[glove.dictionary['trump']]

glove.most_similar("trump",number=10)

glove.most_similar("share",number=10)
